// src/app/business/archivo-clinic/interfaces/Tabla_Arch.ts
export default interface Tabla_arch{
    expediente: string;
    nombre_paciente: string;
    ap_paterno: string;
    ap_materno: string;
    sexo: string;
    edad:number;
    curp: string;
    domicilio:string;
    telefono: string;
    fecha_elab: string;
    elaboro: string;
    ubicacion: string;
    status: number;
}